pyproject.toml
[project]
name = "demo-gemini-agents"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "click>=8.3.1",
    "google-genai>=1.67.0",
    "tenacity>=9.1.4",
]
